+++ pkg/workflow/spec/testutil/json.go
@@ -14,6 +14,10 @@ func JSONSecret(name string) map[string]interface{} {
 	return map[string]interface{}{"$type": "Secret", "name": name}
 }
 
+func JSONConnection(connectionType, name string) map[string]interface{} {
+	return map[string]interface{}{"$type": "Connection", "type": connectionType, "name": name}
+}
+
 func JSONOutput(from, name string) map[string]interface{} {
 	return map[string]interface{}{"$type": "Output", "from": from, "name": name}
 }